minimal code structure of RLCore
