+++ src/ReinforcementLearningCore/Project.toml
@@ -4,19 +4,15 @@ authors = ["Jun Tian <tianjun.cpp@gmail.com>"]
 version = "0.8.11"
 
 [deps]
-AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-ArrayInterface = "4fba245c-0d91-5ea0-9b3e-6abc04ee57a9"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 CircularArrayBuffers = "9de3a189-e0c0-4e15-ba3b-b14b9fb0aec1"
-Compat = "34da2185-b29b-5c13-b0c7-acf172513d20"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ElasticArrays = "fdbdab4c-e67f-52f5-8c3f-e7b388dad3d4"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 Functors = "d9f16b24-f501-4c13-a1f2-28368ffc5196"
-GPUArrays = "0c68f7d7-f131-5f86-a1c3-88cf8149b2d7"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 Markdown = "d6f4376e-aef5-505a-96c1-9c027394607a"
@@ -28,28 +24,20 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Trajectories = "6486599b-a3cd-4e92-a99a-2cea90cc8c3c"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
-Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-AbstractTrees = "0.3"
-Adapt = "2, 3"
-ArrayInterface = "3, 4, 5"
 CUDA = "3.5"
 CircularArrayBuffers = "0.1"
-Compat = "3"
 Distributions = "0.24, 0.25"
-ElasticArrays = "1.2"
 FillArrays = "0.8, 0.9, 0.10, 0.11, 0.12, 0.13"
 Flux = "0.12.9"
 Functors = "0.1, 0.2"
-GPUArrays = "5, 6.0, 7, 8"
 MacroTools = "0.5"
 ProgressMeter = "1.2"
 ReinforcementLearningBase = "0.9"
 Setfield = "0.6, 0.7, 0.8"
 StatsBase = "0.32, 0.33"
 UnicodePlots = "1.3, 2"
-Zygote = "0.5, 0.6"
 julia = "1.6"
 
 [extras]
