+++ src/ReinforcementLearningCore/src/core/core.jl
@@ -1,4 +1,4 @@
-include("hooks.jl")
+include("stages.jl")
 include("stop_conditions.jl")
+include("hooks.jl")
 include("run.jl")
-include("experiment.jl")
