+++ src/ReinforcementLearningCore/src/core/stages.jl
@@ -1,0 +1,21 @@
+export AbstractStage,
+    PreExperimentStage,
+    PostExperimentStage,
+    PreEpisodeStage,
+    PostEpisodeStage,
+    PreActStage,
+    PostActStage
+
+abstract type AbstractStage end
+
+struct PreExperimentStage <: AbstractStage end
+struct PostExperimentStage <: AbstractStage end
+struct PreEpisodeStage <: AbstractStage end
+struct PostEpisodeStage <: AbstractStage end
+struct PreActStage <: AbstractStage end
+struct PostActStage <: AbstractStage end
+
+(p::AbstractPolicy)(::AbstractStage, ::AbstractEnv) = nothing
+(p::AbstractPolicy)(::AbstractStage, ::AbstractEnv, action) = nothing
+
+optimise!(::AbstractPolicy) = nothing
