--- src/ReinforcementLearningCore/src/extensions/extensions.jl
@@ -1,6 +1,0 @@
-include("ArrayInterface.jl")
-include("Flux.jl")
-include("CUDA.jl")
-include("Zygote.jl")
-include("ElasticArrays.jl")
-include("Distributions.jl")
