+++ src/ReinforcementLearningCore/src/utils/utils.jl
@@ -2,3 +2,4 @@ include("basic.jl")
 include("device.jl")
 include("stack_frames.jl")
 include("explorers/explorers.jl")
+include("distributions.jl")
