--- src/ReinforcementLearningCore/test/components/agents.jl
@@ -1,1 +1,0 @@
-@testset "Agent" begin end
