--- src/ReinforcementLearningCore/test/components/components.jl
@@ -1,5 +1,0 @@
-include("processors.jl")
-include("approximators.jl")
-include("explorers.jl")
-include("trajectories.jl")
-include("agents.jl")
